backend/app/models/user_api_key.py
from datetime import datetime
from uuid import UUID, uuid4

from sqlalchemy.dialects.postgresql import UUID as PGUUID
from sqlmodel import Column, Field, SQLModel


class UserApiKey(SQLModel, table=True):
    __tablename__ = "user_api_keys"

    id: UUID = Field(default_factory=uuid4, primary_key=True)
    user_id: UUID = Field(sa_column=Column(PGUUID(as_uuid=True), index=True))
    provider: str
    encrypted_key: str
    is_active: bool = Field(default=True)
    created_at: datetime = Field(default_factory=datetime.utcnow)
