backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Roadmap Viewer Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.30.0",
    "certifi>=2025.11.12",
    "cryptography>=46.0.3",
    "fastapi>=0.121.3",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "langgraph>=0.0.1",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
